src/main/java/bmps/com/dsa/linkedlist/MergeLinkedList.java
package bmps.com.dsa.linkedlist;

public class MergeLinkedList {
    public ListNode<Integer> mergeTwoLists(ListNode<Integer> list1, ListNode<Integer> list2) {
        var dummyNode = new ListNode<Integer>();
        var tail = dummyNode;

        while (list1 != null && list2 != null) {
            if (list1.val < list2.val) {
                tail.next = list1;
                list1 = list1.next;
            } else {
                tail.next = list2;
                list2 = list2.next;
            }

            tail = tail.next;
        }

        if (list1 != null) {
            tail.next = list1;
        } else if (list2 != null) {
            tail.next = list2;
        }

        return dummyNode.next;
    }
}


